requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
autopep8==2.0.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-default-group==1.2.4
click-default-group-wheel==1.2.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==3.0.1
Django==4.2.1
django-allauth==0.57.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-wkhtmltopdf==3.4.0
et-xmlfile==1.1.0
fonttools==4.39.4
freetype-py==2.3.0
greenlet==3.0.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.2
mysql-connector-python==8.0.33
# mysqlclient==2.1.1  # ❌ Eliminado para evitar errores
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.1
Pillow==9.5.0
pluggy==1.3.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.6.0
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.8.0
pylogix==0.8.11
pyModbusTCP==0.2.0
PyMySQL==1.0.3  # ✅ Alternativa ligera y compatible con Railway
pyodbc==4.0.39
pypdf==3.8.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.20
sqlite-fts4==1.0.3
sqlite-utils==3.35
sqlparse==0.4.4
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.2
uvicorn==0.23.2
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
XlsxWriter==3.1.9
zopfli==0.2.2
